pyproject.toml
[tool.poetry]
name = "api-meguiaai"
version = "0.1.0"
description = ""
authors = ["victor-renan <alvesrenanpro@gmail.com>"]
license = "GPL V3"
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.10,<3.11"
numpy = "1.22"
matplotlib = "^3.7.1"
tensorflow = "^2.12.0"
fastapi = "^0.95.2"
uvicorn = "^0.22.0"
python-dotenv = "^1.0.0"
pydantic = "^1.10.7"
psycopg2 = "^2.9.6"
tortoise-orm = "^0.19.3"
peewee = "^3.16.2"
sqlalchemy = ">=1.0,<2.0"


[tool.poetry.group.dev.dependencies]
autopep8 = "^2.0.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
